backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cmake==3.27.4.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
defusedxml==0.7.1
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
easyocr==1.7.1
filelock==3.12.3
flake8==6.1.0
fonttools==4.42.1
idna==3.4
imageio==2.31.3
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
pyclipper==1.3.0.post5
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
sympy==1.12
tifffile==2023.8.30
torch==2.0.1
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
